// src/shared/enums/secretKey.enum.ts
export enum ESecretKey {
  API_KEY = 'API_KEY',
  DATABASE_URL = 'DATABASE_URL',
  DB_USERNAME = 'DB_USERNAME',
  DB_PASSWORD = 'DB_PASSWORD',
  DB_HOST = 'DB_HOST',
  DB_PORT = 'DB_PORT',
  DB_NAME = 'DB_NAME',
  NON_EXISTING_SECRET = 'NON_EXISTING_SECRET',
  MANAGEMENT_API_TOKEN = 'MANAGEMENT_API_TOKEN',
}